extract lots of stuff
